split test suite logically
